* Added "message of the day" (Motd) to ServerInfoMessage. * Added new VersionField derived type of MessageField. * Added unit test for VersionField. * Renamed GuidMessageField to GuidField and StringMessageField to StringField for readability reasons.
git-svn-id: http://localhost/svn/gear/trunk@107 f7f8973e-caa7-11de-a1cf-4bfa7bdbae30

+++ src/Gear/Net/Message.cs
@@ -63,7 +63,7 @@ namespace Gear.Net
         }
         #endregion
         #region Methods
-        public abstract MessageField GetField(MessageFieldId id, byte tag);
+        public abstract MessageField GetField(MessageFieldId id, short tag);
         #endregion
     }
 }
